Moved mac_style check to OT profile
