+++ Lib/fontbakery/profiles/opentype.py
@@ -76,6 +76,7 @@ OPENTYPE_PROFILE_CHECKS = [
     'com.google.fonts/check/layout_valid_script_tags',
     'com.google.fonts/check/layout_valid_language_tags',
     'com.google.fonts/check/italic_angle',
+    'com.google.fonts/check/mac_style',
     'com.adobe.fonts/check/varfont/valid_axis_nameid',
     'com.adobe.fonts/check/varfont/valid_subfamily_nameid',
     'com.adobe.fonts/check/varfont/valid_postscript_nameid',
