+++ tests/profiles/googlefonts_test.py
@@ -2321,38 +2321,6 @@ def test_check_slant_direction():
                            FAIL, 'positive-value-for-clockwise-lean')
 
 
-def test_check_mac_style():
-    """ Checking head.macStyle value. """
-    check = CheckTester(googlefonts_profile,
-                        "com.google.fonts/check/mac_style")
-    from fontbakery.constants import MacStyle
-
-    ttFont = TTFont(TEST_FILE("cabin/Cabin-Regular.ttf"))
-
-    # macStyle-value, style, expected
-    test_cases = [
-        [0, "Thin", PASS],
-        [0, "Bold", "bad-BOLD"],
-        [0, "Italic", "bad-ITALIC"],
-        [MacStyle.ITALIC, "Italic", PASS],
-        [MacStyle.ITALIC, "Thin", "bad-ITALIC"],
-        [MacStyle.BOLD, "Bold", PASS],
-        [MacStyle.BOLD, "Thin", "bad-BOLD"],
-        [MacStyle.BOLD | MacStyle.ITALIC, "BoldItalic", PASS]
-    ]
-
-    for macStyle_value, style, expected in test_cases:
-        ttFont["head"].macStyle = macStyle_value
-
-        if expected == PASS:
-            assert_PASS(check(ttFont, {"style": style}),
-                        'with macStyle:{macStyle_value} style:{style}...')
-        else:
-            assert_results_contain(check(ttFont, {"style": style}),
-                                   FAIL, expected,
-                                   f"with macStyle:{macStyle_value} style:{style}...")
-
-
 # FIXME!
 # GFonts hosted Cabin files seem to have changed in ways
 # that break some of the assumptions in the code-test below.
